Refactored the ModuleForms
